Add myself to gem authors
I'm very doubting in it. I'm contributor and kind of maintainer, yes,
but I've done not too much impact in comparison with the original author.

But I'd like to leave "actual" contact info.
It's not necessary to do it in the gem specs,
but it can be uselful there for somebody.